refactor + feat: Supabase auth, login page
