+++ package.json
@@ -17,6 +17,9 @@
     "@mui/material-nextjs": "^5.15.6",
     "@mui/styled-engine-sc": "^6.0.0-alpha.9",
     "@reduxjs/toolkit": "^2.0.1",
+    "@supabase/auth-helpers-nextjs": "^0.9.0",
+    "@supabase/ssr": "^0.1.0",
+    "@supabase/supabase-js": "^2.39.3",
     "@vercel/postgres": "^0.5.1",
     "mongodb": "^6.3.0",
     "mongoose": "^8.0.3",
