+++ src/app/(user)/[slug]/page.tsx
