+++ src/app/(user)/[slug]/seat.tsx
