+++ src/app/context/StoreProvider.tsx
@@ -1,0 +1,19 @@
+"use client";
+import { AppStore, makeStore } from "@/src/utils/store/store";
+import { useRef } from "react";
+import { Provider } from "react-redux";
+
+
+export default function StoreProvider({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const storeRef = useRef<AppStore>();
+  if (!storeRef.current) {
+    // Create the store instance the first time this renders
+    storeRef.current = makeStore();
+  }
+
+  return <Provider store={storeRef.current}>{children}</Provider>;
+}
