+++ src/app/globals.css
@@ -1,0 +1,70 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+
+
+:root {
+  --foreground-rgb: 0, 0, 0;
+  --background-start-rgb: 214, 219, 220;
+  --background-end-rgb: 255, 255, 255;
+}
+
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    --foreground-rgb: 255, 255, 255;
+    --background-start-rgb: 0, 0, 0;
+    --background-end-rgb: 0, 0, 0;
+  }
+}
+
+body {
+  color: rgb(var(--foreground-rgb));
+  background-color: #f4f4f4;
+ 
+}
+.border-plane{
+  border-top-right-radius: 8rem 20rem;
+  border-top-left-radius: 8rem 20rem;
+  border-bottom-right-radius: 5rem 8rem;
+  border-bottom-left-radius: 5rem 8rem;
+  padding: 17rem 0.5rem 10rem ;
+  border: 0.3rem solid #d2d6d9; 
+  margin-top: -11rem; 
+}
+@media screen and (min-width: 640px){
+  .border-plane{
+    border-top-right-radius: 12rem 20rem;
+    border-top-left-radius: 12rem 20rem;
+    padding: 17rem 2rem 10rem;
+  }
+  .boxed-border{
+    background-image: linear-gradient(to bottom,var(--light-base, #fff),var(--bg-blue, #f2f8fc) 1%);
+    border: 1px solid #c3c7cc;
+    border-top: none;
+  }
+  
+}
+
+
+.bg-plane{
+  
+  background-image:  url("/airplane.png");
+  background-repeat: no-repeat;
+  background-clip: border-box;
+  background-position: center bottom;
+
+
+}
+
+
+
+.text-stroke-white{
+  mix-blend-mode: difference;
+  color:black;
+  filter:invert(100%)
+}
+.text-stroke-black{
+  mix-blend-mode: difference;
+}
