+++ src/app/layout.tsx
@@ -1,0 +1,28 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { AppRouterCacheProvider } from "@mui/material-nextjs/v13-appRouter";
+import "./globals.css";
+import StoreProvider from "./context/StoreProvider";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <html lang="en">
+      <StoreProvider>
+        <body className={inter.className}>
+          <AppRouterCacheProvider>{children}</AppRouterCacheProvider>
+        </body>
+      </StoreProvider>
+    </html>
+  );
+}
