+++ src/components/(admin)/navBar.tsx
@@ -1,0 +1,9 @@
+import { Button } from "@mui/material";
+
+export default function Navbar(){
+    return(
+        <nav>
+            <Button>Logout</Button>
+        </nav>
+    )
+}
